Move generic utils from __init__.py to basic.py.
